+++ cabi/include/semaphore.h
@@ -9,6 +9,16 @@
 #include <stdlib.h>
 
 /**
+ * Controls the globbing behaviors.
+ */
+enum GlobFlags {
+  GLOB_FLAGS_DOUBLE_STAR = 1,
+  GLOB_FLAGS_CASE_INSENSITIVE = 2,
+  GLOB_FLAGS_PATH_NORMALIZE = 4,
+};
+typedef uint32_t GlobFlags;
+
+/**
  * Represents all possible error codes
  */
 enum SemaphoreErrorCode {
@@ -135,6 +145,8 @@ SemaphoreUuid semaphore_get_register_response_relay_id(const SemaphoreBuf *data)
  */
 void semaphore_init(void);
 
+bool semaphore_is_glob_match(const SemaphoreBuf *value, const SemaphoreStr *pat, GlobFlags flags);
+
 /**
  * Frees a public key.
  */
@@ -230,6 +242,11 @@ bool semaphore_uuid_is_nil(const SemaphoreUuid *uuid);
  * `semaphore_cstr_free`.
  */
 SemaphoreStr semaphore_uuid_to_str(const SemaphoreUuid *uuid);
+
+/**
+ * Returns a list of all valid platform identifiers.
+ */
+const SemaphoreStr *semaphore_valid_platforms(uintptr_t *size_out);
 
 /**
  * Validates a register response.