+++ common/src/lib.rs
@@ -9,6 +9,7 @@ pub mod metrics;
 
 mod auth;
 mod config;
+mod glob;
 mod log;
 mod retry;
 mod types;
@@ -17,6 +18,7 @@ mod utils;
 
 pub use crate::auth::*;
 pub use crate::config::*;
+pub use crate::glob::*;
 pub use crate::log::*;
 pub use crate::retry::*;
 pub use crate::types::*;