+++ general/src/protocol/logentry.rs
@@ -1,5 +1,5 @@
 use crate::protocol::JsonLenientString;
-use crate::types::{Annotated, Error, FromValue, Object, Value};
+use crate::types::{Annotated, Error, FromValue, Meta, Object, Value};
 
 /// A log entry message.
 ///
@@ -74,7 +74,7 @@ impl FromValue for LogEntry {
             // The next two cases handle the legacy top-level `message` attribute, which was sent as
             // literal string, false (which should be ignored) or even as deep JSON object. Sentry
             // historically JSONified this field.
-            Annotated(Some(Value::Bool(false)), _) => Annotated(None, Default::default()),
+            Annotated(Some(Value::Bool(false)), _) => Annotated(None, Meta::default()),
             x => Annotated::new(LogEntry {
                 formatted: JsonLenientString::from_value(x)
                     .map_value(JsonLenientString::into_inner),