+++ server/src/actors/project.rs
@@ -1,6 +1,5 @@
 use std::borrow::Cow;
-use std::collections::hash_map::Entry;
-use std::collections::HashMap;
+use std::collections::{hash_map::Entry, HashMap, VecDeque};
 use std::ffi::OsStr;
 use std::fs;
 use std::io;
@@ -9,7 +8,7 @@ use std::mem;
 use std::path::Path;
 use std::sync::Arc;
 use std::thread;
-use std::time::{Duration, Instant, SystemTime};
+use std::time::{Duration, Instant};
 
 use actix::fut;
 use actix::prelude::*;
@@ -32,7 +31,7 @@ use crate::actors::controller::{Controller, Shutdown, Subscribe, TimeoutError};
 use crate::actors::outcome::DiscardReason;
 use crate::actors::upstream::{SendQuery, UpstreamQuery, UpstreamRelay};
 use crate::extractors::EventMeta;
-use crate::utils::{self, One, Response, SyncActorFuture, SyncHandle};
+use crate::utils::{self, ErrorBoundary, One, Response, SyncActorFuture, SyncHandle};
 
 #[derive(Fail, Debug)]
 pub enum ProjectError {
@@ -95,7 +94,7 @@ impl Project {
         if let Some(ref state) = self.state {
             // In case the state is fetched from a local file, don't use own caching logic. Rely on
             // `ProjectCache#local_states` for caching.
-            if !self.is_local && !state.outdated(&self.config) {
+            if !self.is_local && !state.outdated(self.id, &self.config) {
                 return Response::ok(state.clone());
             }
         }
@@ -277,6 +276,10 @@ pub struct ProjectState {
     /// The organization id.
     #[serde(default)]
     pub organization_id: Option<u64>,
+
+    /// True if this project state was fetched but incompatible with this Relay.
+    #[serde(skip, default)]
+    pub invalid: bool,
 }
 
 impl ProjectState {
@@ -288,9 +291,10 @@ impl ProjectState {
             disabled: true,
             public_keys: Vec::new(),
             slug: None,
-            config: Default::default(),
+            config: ProjectConfig::default(),
             rev: None,
             organization_id: None,
+            invalid: false,
         }
     }
 
@@ -300,6 +304,13 @@ impl ProjectState {
     pub fn allowed() -> Self {
         let mut state = ProjectState::missing();
         state.disabled = false;
+        state
+    }
+
+    /// Project state for a deserialization error.
+    pub fn err() -> Self {
+        let mut state = ProjectState::missing();
+        state.invalid = true;
         state
     }
 
@@ -332,15 +343,43 @@ impl ProjectState {
         self.disabled
     }
 
+    /// Returns `true` if the project state obtained from the upstream could not be parsed. This
+    /// results in events being dropped similar to disabled states, but can provide separate
+    /// metrics.
+    pub fn invalid(&self) -> bool {
+        self.invalid
+    }
+
     /// Returns whether this state is outdated and needs to be refetched.
-    pub fn outdated(&self, config: &Config) -> bool {
-        SystemTime::from(self.last_fetch)
-            .elapsed()
-            .map(|e| match self.slug {
-                Some(_) => e > config.project_cache_expiry(),
-                None => e > config.cache_miss_expiry(),
-            })
-            .unwrap_or(false)
+    pub fn outdated(&self, project_id: ProjectId, config: &Config) -> bool {
+        let expiry = match self.slug {
+            Some(_) => config.project_cache_expiry(),
+            None => config.cache_miss_expiry(),
+        };
+
+        // Project state updates are aligned to a fixed grid based on the expiry interval. By
+        // default, that's a grid of 1 minute intervals for invalid projects, and a grid of 5
+        // minutes for existing projects (cache hits). The exception to this is when a project is
+        // seen for the first time, where it is fetched immediately.
+        let window = expiry.as_secs();
+
+        // To spread out project state updates more evenly, they are shifted deterministically
+        // within the grid window. A 5 minute interval results in 300 theoretical slots that can be
+        // chosen for each project based on its project id.
+        let project_shift = project_id % window;
+
+        // Based on the last fetch, compute the timestamp of the next fetch. The time stamp is
+        // shifted by the project shift to move the grid accordingly. Note that if the remainder is
+        // zero, the next fetch is one full window ahead to avoid instant reloading.
+        let last_fetch = self.last_fetch.timestamp() as u64;
+        let remainder = (last_fetch - project_shift) % window;
+        let next_fetch = last_fetch + (window - remainder);
+
+        // See the below assertion for constraints on the next fetch time.
+        debug_assert!(next_fetch > last_fetch && next_fetch <= last_fetch + window);
+
+        // A project state counts as outdated when the time of the next fetch has passed.
+        Utc::now().timestamp() as u64 >= next_fetch
     }
 
     /// Returns the project config.
@@ -366,13 +405,18 @@ impl ProjectState {
     }
 
     /// Determines whether the given event should be accepted or dropped.
-    pub fn get_event_action(&self, meta: &EventMeta, config: &Config) -> EventAction {
+    pub fn get_event_action(
+        &self,
+        project_id: ProjectId,
+        meta: &EventMeta,
+        config: &Config,
+    ) -> EventAction {
         // Try to verify the request origin with the project config.
         if !self.is_valid_origin(meta.origin()) {
-            return EventAction::Discard(DiscardReason::ProjectId);
+            return EventAction::Discard(DiscardReason::Cors);
         }
 
-        if self.outdated(config) {
+        if self.outdated(project_id, config) {
             // if the state is out of date, we proceed as if it was still up to date. The
             // upstream relay (or sentry) will still filter events.
 
@@ -382,10 +426,16 @@ impl ProjectState {
             // available in the meanwhile.
             match self.get_public_key_status(meta.auth().public_key()) {
                 PublicKeyStatus::Enabled => EventAction::Accept,
-                PublicKeyStatus::Disabled => EventAction::Discard(DiscardReason::AuthClient),
+                PublicKeyStatus::Disabled => EventAction::Discard(DiscardReason::ProjectId),
                 PublicKeyStatus::Unknown => EventAction::Accept,
             }
         } else {
+            // if we recorded an invalid project state response from the upstream (i.e. parsing
+            // failed), discard the event with a s
+            if self.invalid() {
+                return EventAction::Discard(DiscardReason::ProjectState);
+            }
+
             // only drop events if we know for sure the project is disabled.
             if self.disabled() {
                 return EventAction::Discard(DiscardReason::ProjectId);
@@ -397,10 +447,10 @@ impl ProjectState {
             // events are sent to the upstream.
             match self.get_public_key_status(meta.auth().public_key()) {
                 PublicKeyStatus::Enabled => EventAction::Accept,
-                PublicKeyStatus::Disabled => EventAction::Discard(DiscardReason::AuthClient),
+                PublicKeyStatus::Disabled => EventAction::Discard(DiscardReason::ProjectId),
                 PublicKeyStatus::Unknown => match config.relay_mode() {
                     RelayMode::Proxy => EventAction::Accept,
-                    _ => EventAction::Discard(DiscardReason::AuthClient),
+                    _ => EventAction::Discard(DiscardReason::ProjectId),
                 },
             }
         }
@@ -602,6 +652,8 @@ impl Handler<GetEventAction> for Project {
     type Result = Response<EventAction, ProjectError>;
 
     fn handle(&mut self, message: GetEventAction, context: &mut Self::Context) -> Self::Result {
+        let project_id = self.id;
+
         // Check for an eventual rate limit. Note that we need to check for all variants of
         // RateLimitScope, otherwise we might miss a rate limit.
         let rate_limit = RateLimitScope::iter_variants(&message.meta)
@@ -626,13 +678,13 @@ impl Handler<GetEventAction> for Project {
             // This will return synchronously if the state is still cached.
             let config = self.config.clone();
             self.get_or_fetch_state(context)
-                .map(move |state| state.get_event_action(&message.meta, &config))
+                .map(move |state| state.get_event_action(project_id, &message.meta, &config))
         } else {
             // Fetching is not permitted (as part of the store request). In case the state is not
             // cached, assume that the event can be accepted. The EventManager will later fetch the
             // project state and reevaluate the event action.
             Response::ok(self.state().map_or(EventAction::Accept, |state| {
-                state.get_event_action(&message.meta, &self.config)
+                state.get_event_action(project_id, &message.meta, &self.config)
             }))
         }
     }
@@ -705,7 +757,7 @@ pub struct GetProjectStates {
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct GetProjectStatesResponse {
-    pub configs: HashMap<ProjectId, Option<ProjectState>>,
+    pub configs: HashMap<ProjectId, ErrorBoundary<Option<ProjectState>>>,
 }
 
 impl UpstreamQuery for GetProjectStates {
@@ -720,6 +772,21 @@ impl UpstreamQuery for GetProjectStates {
     }
 }
 
+#[derive(Clone, Copy, Debug)]
+struct ProjectUpdate {
+    project_id: ProjectId,
+    instant: Instant,
+}
+
+impl ProjectUpdate {
+    pub fn new(project_id: ProjectId) -> Self {
+        ProjectUpdate {
+            project_id,
+            instant: Instant::now(),
+        }
+    }
+}
+
 pub struct ProjectCache {
     backoff: RetryBackoff,
     config: Arc<Config>,
@@ -727,6 +794,7 @@ pub struct ProjectCache {
     projects: HashMap<ProjectId, Addr<Project>>,
     local_states: HashMap<ProjectId, Arc<ProjectState>>,
     state_channels: HashMap<ProjectId, oneshot::Sender<ProjectState>>,
+    updates: VecDeque<ProjectUpdate>,
     shutdown: SyncHandle,
 }
 
@@ -739,6 +807,7 @@ impl ProjectCache {
             projects: HashMap::new(),
             local_states: HashMap::new(),
             state_channels: HashMap::new(),
+            updates: VecDeque::new(),
             shutdown: SyncHandle::new(),
         }
     }
@@ -775,18 +844,51 @@ impl ProjectCache {
             self.backoff.attempt(),
         );
 
+        let eviction_start = Instant::now();
+
+        // Remove outdated projects that are not being refreshed from the cache. If the project is
+        // being updated now, also remove its update entry from the queue, since we will be
+        // inserting a new timestamp at the end (see `extend`).
+        let eviction_threshold = eviction_start - 2 * self.config.project_cache_expiry();
+        while let Some(update) = self.updates.get(0) {
+            if update.instant > eviction_threshold {
+                break;
+            }
+
+            if !channels.contains_key(&update.project_id) {
+                self.projects.remove(&update.project_id);
+            }
+
+            self.updates.pop_front();
+        }
+
+        // The remaining projects are not outdated anymore. Still, clean them from the queue to
+        // reinsert them at the end, as they are now receiving an updated timestamp. Then,
+        // batch-insert all new projects with the new timestamp.
+        self.updates
+            .retain(|update| !channels.contains_key(&update.project_id));
+        self.updates
+            .extend(channels.keys().copied().map(ProjectUpdate::new));
+
+        metric!(timer("project_state.eviction.duration") = eviction_start.elapsed());
+
         let request = GetProjectStates {
-            projects: channels.keys().cloned().collect(),
+            projects: channels.keys().copied().collect(),
             #[cfg(feature = "processing")]
             full_config: self.config.processing_enabled(),
         };
+
         // count number of http requests for project states
         metric!(counter("project_state.request") += 1);
+        let request_start = Instant::now();
+
         self.upstream
             .send(SendQuery(request))
             .map_err(ProjectError::ScheduleFailed)
             .into_actor(self)
-            .and_then(|response, slf, ctx| {
+            .and_then(move |response, slf, ctx| {
+                metric!(timer("project_state.request.duration") = request_start.elapsed());
+
                 match response {
                     Ok(mut response) => {
                         slf.backoff.reset();
@@ -797,7 +899,8 @@ impl ProjectCache {
                             let state = response
                                 .configs
                                 .remove(&id)
-                                .unwrap_or(None)
+                                .unwrap_or(ErrorBoundary::Ok(None))
+                                .unwrap_or_else(|| Some(ProjectState::err()))
                                 .unwrap_or_else(ProjectState::missing);
 
                             channel.send(state).ok();
@@ -957,16 +1060,15 @@ impl Handler<GetProject> for ProjectCache {
         let config = self.config.clone();
         metric!(histogram("project_cache.size") = self.projects.len() as u64);
         match self.projects.entry(message.id) {
-            Entry::Occupied(value) => {
+            Entry::Occupied(entry) => {
                 metric!(counter("project_cache.hit") += 1);
-                value.get().clone()
+                entry.get().clone()
             }
-            Entry::Vacant(_) => {
+            Entry::Vacant(entry) => {
                 metric!(counter("project_cache.miss") += 1);
-                let new_val = Project::new(message.id, config, context.address()).start();
-                let ret_val = new_val.clone();
-                self.projects.insert(message.id, new_val);
-                ret_val
+                let project = Project::new(message.id, config, context.address()).start();
+                entry.insert(project.clone());
+                project
             }
         }
     }